Added wrapper for axios

+++ client/app/redux/modules/auth/actions.ts
@@ -7,46 +7,15 @@ import history from 'components/containers/history';
 
 const axiosRequest: any = axios;
 
-export type Login = (Email: string, Password: string) => (dispatch: Dispatch) => void;
-export const login: Login = (Email: string, Password: string) => (dispatch: Dispatch) => {
-  return axiosRequest
-    .post('/api/auth', { Email, Password })
-    .then((res: axios.AxiosResponse<ILogin>) => {
-      const { AccessToken, RefreshToken } = res.data;
-      const User = JSON.stringify(res.data.Data);
+export type Login = (
+  Email: string,
+  Password: string,
+) => (dispatch: Dispatch) => axios.AxiosPromise;
 
-      localStorage.setItem('AccessToken', AccessToken);
-      localStorage.setItem('RefreshToken', RefreshToken);
-      localStorage.setItem('User', User);
-
-      dispatch({
-        type: AUTH_CONSTANTS.LOGIN,
-        payload: res.data,
-      });
-    })
-    .catch((err: axios.AxiosError) => {
-      console.error(err);
-      throw  err;
-    });
-};
-
-export type VerifyCredentials = (dispatch: Dispatch) => void;
-export const verifyCredentials: VerifyCredentials = (dispatch: Dispatch) => {
-  const accessToken = localStorage.getItem('AccessToken');
-  const refreshToken = localStorage.getItem('RefreshToken');
-  const currentUser = JSON.parse(localStorage.getItem('User'));
-
-  if (accessToken && refreshToken && currentUser) {
-    dispatch({
-      type: AUTH_CONSTANTS.LOGIN,
-      payload: {
-        AccessToken: accessToken,
-        RefreshToken: refreshToken,
-        Data: currentUser,
-      },
-    });
-    axiosRequest
-      .patch('/api/auth/verify-credentials', { accessToken, refreshToken })
+export function login(Email: string, Password: string) {
+  return (dispatch: Dispatch) => {
+    return axiosRequest
+      .post('/api/auth', { Email, Password })
       .then((res: axios.AxiosResponse<ILogin>) => {
         const { AccessToken, RefreshToken } = res.data;
         const User = JSON.stringify(res.data.Data);
@@ -61,22 +30,63 @@ export const verifyCredentials: VerifyCredentials = (dispatch: Dispatch) => {
         });
       })
       .catch((err: axios.AxiosError) => {
-        if (err.response.status === 401) {
-          localStorage.clear();
-          dispatch({
-            type: AUTH_CONSTANTS.LOGOUT,
-          });
-        } else {
-          console.error(err);
-          throw  err;
-        }
+        console.error(err);
+        throw err;
       });
+  };
+}
+
+export type VerifyCredentials = (dispatch: Dispatch) => void;
+export const verifyCredentials: VerifyCredentials = async (
+  dispatch: Dispatch,
+) => {
+  const accessToken = localStorage.getItem('AccessToken');
+  const refreshToken = localStorage.getItem('RefreshToken');
+  const currentUser = JSON.parse(localStorage.getItem('User'));
+
+  if (accessToken && refreshToken && currentUser) {
+    dispatch({
+      type: AUTH_CONSTANTS.LOGIN,
+      payload: {
+        AccessToken: accessToken,
+        RefreshToken: refreshToken,
+        Data: currentUser,
+      },
+    });
+    try {
+      const res = await axiosRequest.patch('/api/auth/verify-credentials', {
+        accessToken,
+        refreshToken,
+      });
+      const { AccessToken, RefreshToken } = res.data;
+      const User = JSON.stringify(res.data.Data);
+
+      localStorage.setItem('AccessToken', AccessToken);
+      localStorage.setItem('RefreshToken', RefreshToken);
+      localStorage.setItem('User', User);
+
+      dispatch({
+        type: AUTH_CONSTANTS.LOGIN,
+        payload: res.data,
+      });
+    } catch (err) {
+      if (err.response.status === 401) {
+        localStorage.clear();
+        dispatch({
+          type: AUTH_CONSTANTS.LOGOUT,
+        });
+      } else {
+        console.error(err);
+        throw err;
+      }
+    }
   }
 };
 
 export type SignUp = (registeredUser: IRegisteredUser) => (dispatch: Dispatch) => void;
 export const signUp: SignUp = (registeredUser: IRegisteredUser) => (dispatch: Dispatch) => {
-  return axiosRequest.post('/api/users', registeredUser)
+  return axiosRequest
+    .post('/api/users', registeredUser)
     .then((res: axios.AxiosResponse<void>) => {
       if (res.status === 201) {
         history.push('/success');
@@ -86,22 +96,31 @@ export const signUp: SignUp = (registeredUser: IRegisteredUser) => (dispatch: Di
       if (err.response.status === 400) {
         dispatch({
           type: APP_ACTIONS.ADD_NOTIFICATION,
-          payload: { status: err.response.status, message: 'User with the same email already exist' },
+          payload: {
+            status: err.response.status,
+            message: 'User with the same email already exist',
+          },
         });
       } else {
         dispatch({
           type: APP_ACTIONS.ADD_NOTIFICATION,
-          payload: { status: err.response.status, message: err.response.statusText },
+          payload: {
+            status: err.response.status,
+            message: err.response.statusText,
+          },
         });
       }
       console.error(err);
-      throw  err;
+      throw err;
     });
 };
 
 export type Logout = (refreshToken: string) => (dispatch: Dispatch) => void;
-export const logout: Logout = (refreshToken: string) => (dispatch: Dispatch) => {
-  return axiosRequest.delete(`/api/auth/${refreshToken}`)
+export const logout: Logout = (refreshToken: string) => (
+  dispatch: Dispatch,
+) => {
+  return axiosRequest
+    .delete(`/api/auth/${refreshToken}`)
     .then(() => {
       // TODO: import logout function from helper
       localStorage.clear();
@@ -121,7 +140,7 @@ export const logout: Logout = (refreshToken: string) => (dispatch: Dispatch) => 
         history.push('/error/server-error"');
       } else {
         console.error(err);
-        throw  err;
+        throw err;
       }
     });
 };
